Code cleanup

+++ include/systems/CameraMovementSystem.h
@@ -18,15 +18,15 @@ public:
 
             const auto posX = transform.position.x;
             const auto posY = transform.position.y;
-            const auto offsetX = sprite.rect.w * transform.scale.x * .5f;
-            const auto offsetY = sprite.rect.h * transform.scale.y * .5f;
-            const auto widthShift = (camera.w >> 1) - offsetX;
-            const auto heightShift = (camera.h >> 1) - offsetY;
-            if (posX > widthShift && posX < Game::mapWidth - widthShift - offsetX * 2) {
-                camera.x = posX - widthShift;
+            const auto offsetX = static_cast<float>(sprite.rect.w) * transform.scale.x * .5f;
+            const auto offsetY = static_cast<float>(sprite.rect.h) * transform.scale.y * .5f;
+            const auto widthShift = static_cast<float>(camera.w >> 1) - offsetX;
+            const auto heightShift = static_cast<float>(camera.h >> 1) - offsetY;
+            if (posX > widthShift && posX < static_cast<float>(Game::mapWidth) - widthShift - offsetX * 2) {
+                camera.x = static_cast<int>(posX - widthShift);
             }
-            if (posY > heightShift && posY < Game::mapHeight - heightShift - offsetY * 2) {
-                camera.y = posY - heightShift;
+            if (posY > heightShift && posY < static_cast<float>(Game::mapHeight) - heightShift - offsetY * 2) {
+                camera.y = static_cast<int>(posY - heightShift);
             }
             camera.x = camera.x < 0 ? 0 : camera.x > camera.w ? camera.w : camera.x;
             camera.y = camera.y < 0 ? 0 : camera.y > camera.h ? camera.h : camera.y;
